[REF] Refatoração da estrutura dos arquivos
